+++ caten_worship/models/users.py
@@ -14,9 +14,10 @@ db = SQLAlchemy(app)
 class User(db.Model):
     __tablename__ = "users"
     id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(64), unique=True, nullable=False)
-    email = db.Column(db.String(128), unique=True, nullable=False)
-    password_hash = db.Column(db.String(128), nullable=False)
+    username = db.Column(db.String(24), unique=True, nullable=False)
+    email = db.Column(db.String(64), unique=True, nullable=False)
+    password_hash = db.Column(db.String(64), nullable=False)
+    displayname = db.Column(db.String(16), nullable=False)
     activated = db.Column(db.Boolean, default=False)
 
     @property