+++ caten_worship/__init__.py
@@ -2,14 +2,13 @@
 
 # -*- coding: utf-8 -*-
 
-import os
 from flask import Flask
 
 
-def create_app():
+def create_app(conf):
     from . import routes, models, db
     app = Flask(__name__)
-    app.config.from_object(os.environ.get("APP_SETTING"))
+    app.config.from_object(conf)
     routes.init_app(app)
     models.init_app(app)
     db.init_app(app)