+++ config.py
@@ -33,5 +33,7 @@ class Config(object):
     class Testing(Default):
         ENV = "testing"
         TESTING = True
+        SQLALCHEMY_TRACK_MODIFICATIONS = False
+        SQLALCHEMY_DATABASE_URI = os.environ.get("DATABASE_URL_FOR_TESTING")
 
 del os