+++ caten_worship/__init__.py
@@ -1,16 +1,22 @@
 # caten_worship/__init__.py
 
-# -*- coding: utf-8 -*-
+
+import os
 
 from flask import Flask
 
+config = os.environ.get("APP_SETTING")
 
-def create_app(conf):
-    from . import routes, models, db
+
+def create_app():
+
     app = Flask(__name__)
-    app.config.from_object(conf)
-    routes.init_app(app)
+    app.config.from_object(config)
+
+    from . import models
     models.init_app(app)
-    db.init_app(app)
+
+    from . import routes
+    routes.init_app(app)
 
     return app