+++ caten_worship/models/__init__.py
@@ -1,5 +1,6 @@
 # models/__init__.py
 
+from .base import db
 from .songs import songsDB
 from .dbx import get_dbx
 from .users import User
@@ -11,3 +12,5 @@ from .login_managers import login_manager
 def init_app(app):
     mail.init_app(app)
     login_manager.init_app(app)
+    db.create_all()
+    db.init_app(app)