+++ run.py
@@ -1,10 +1,9 @@
 # run.py
 
-import os
 
 from caten_worship import create_app
 
-app = create_app(os.environ.get("APP_SETTING"))
+app = create_app()
 
 if __name__ == "__main__":
     app.run()