+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### Next Todo
 
+- UnitTest
 - db migration
 - 會員資料: 編輯、管理